+++ spring-web-reactive/src/main/java/org/springframework/web/reactive/function/client/WebClientStrategies.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.web.client.reactive;
+package org.springframework.web.reactive.function.client;
 
 import java.util.function.Supplier;
 import java.util.stream.Stream;
