+++ spring-web-reactive/src/main/java/org/springframework/web/reactive/function/client/package-info.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright 2002-2016 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Provides a reactive {@link org.springframework.web.reactive.function.client.WebClient}
+ * that builds on top of the
+ * {@code org.springframework.http.client.reactive} reactive HTTP adapter layer.
+ */
+package org.springframework.web.reactive.function.client;
