+++ spring-web-reactive/src/test/java/org/springframework/web/reactive/function/client/DefaultClientResponseTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.web.client.reactive;
+package org.springframework.web.reactive.function.client;
 
 import java.net.InetSocketAddress;
 import java.nio.ByteBuffer;
