--- spring-web/src/main/java/org/springframework/web/client/reactive/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Provides a reactive {@link org.springframework.web.client.reactive.WebClient}
- * that builds on top of the
- * {@code org.springframework.http.client.reactive} reactive HTTP adapter layer.
- */
-package org.springframework.web.client.reactive;
